changet tipe ID

+++ src/com/exam/app/CSVReader.java
@@ -4,7 +4,7 @@ import java.util.HashMap;
 
 public class CSVReader implements Reader {
     @Override
-    public HashMap<Integer, Client> readData() {
+    public HashMap<Long, Client> readData() {
         return null;
     }
 }
